+++ supvisors/test/scripts/start_all.sh
@@ -18,7 +18,7 @@ do
       ping -c 1 $host 2>&1 >/dev/null && ssh $host "cd $TEST_DIR
       export DISPLAY=:0
       export IDENTIFIER=supv-03
-      supervisord -i supv-03 -c etc/supervisord_alt.conf"
+      supervisord -c etc/supervisord_alt.conf"
     else
       echo "start Supervisor on $host"
       ping -c 1 $host 2>&1 >/dev/null && ssh $host "cd $TEST_DIR ; rm -rf log/*