+++ supvisors/tests/conftest.py
@@ -73,7 +73,7 @@ def supervisor():
 
 @pytest.fixture
 def supvisors(mocker, supervisor, options):
-    mocker.patch('supvisors.supvisorsmapper.get_addresses', side_effect=lambda x, y: (x, [], [x]))
+    mocker.patch('supvisors.supvisorsmapper.get_addresses', side_effect=lambda x, y: (x, [x], [x]))
     return MockedSupvisors(supervisor, options)